added readable path and its handling
now book can have pdf inside

+++ libralab-api/src/controller/bookController.js
@@ -3,6 +3,10 @@ import * as jwtMiddleware from '../middleware/JWTMiddleware.js'
 import * as path from 'path'
 import fs from 'fs'
 
+
+//to-do -> fix book controller with both pdf and image file 
+
+
 export async function postBook(req, res){
     try {
         const authHeader = req.cookies.jwt;
@@ -19,17 +23,21 @@ export async function postBook(req, res){
             return
         }
         
-        if (!req.file) {
+        if (!req.files || (!req.files.coverfile && req.files.readablefile)) {
             return res.status(400).json({
               success: false,
-              message: 'No cover file upload',
+              message: 'No cover file or readable file upload',
             });
           }
+
+        const coverFile = req.files.coverfile[0];
+        const readableFile = req.files.readablefile[0];
         
         const bookData = {
             ...req.body, 
-            ID_Author: Token.ID_Author, 
-            cover_path: req.file.filename
+            ID_Author: Token.ID_Author,
+            cover_path : coverFile.filename || null,
+            readable_path : readableFile.filename || null
         };
         
         const result = await bookModel.postBookDb(bookData);
@@ -47,6 +55,17 @@ export async function postBook(req, res){
                 console.log(`Image deleted: ${imageFolderPath}`);
             }
 
+            //Rollback readable update
+            const { readable_path } = bookData;
+
+            const readableFolderPath = path.join(process.cwd(), '/libralab-api/media/readable/book', readable_path);
+
+            if (fs.existsSync(readableFolderPath)) {
+
+                await fs.promises.unlink(readableFolderPath); 
+                console.log(`Readable deleted: ${readableFolderPath}`);
+            }
+
             res.status(500).json({
                 message: 'Error sending book Information'});
             return
@@ -65,6 +84,17 @@ export async function postBook(req, res){
                 console.log(`Image deleted: ${imageFolderPath}`);
             }
 
+            //Rollback readable update
+            const { readable_path } = bookData;
+
+            const readableFolderPath = path.join(process.cwd(), '/libralab-api/media/readable/book', readable_path);
+
+            if (fs.existsSync(readableFolderPath)) {
+
+                await fs.promises.unlink(readableFolderPath); 
+                console.log(`Readable deleted: ${readableFolderPath}`);
+            }
+
             res.status(503).json({
                 message: 'Error sending book Information'});
             return
@@ -78,7 +108,7 @@ export async function postBook(req, res){
             message: 'Internal server error' })
         console.log(error);
     }
-}
+}//still broken
 
 export async function putBookByBookId(req,res){
     try {const authHeader = req.cookies.jwt;
@@ -95,39 +125,33 @@ export async function putBookByBookId(req,res){
             return
         }
         
-        if (!req.file) {
-            return res.status(400).json({
+        if (!req.files || (!req.files.coverfile && req.files.readablefile)) {
+            res.status(400).json({
               success: false,
-              message: 'No new cover file upload',
+              message: 'No cover file or readable file upload',
             });
+            return
+
+            
           }
 
-        // Path to the file on the server
-        // const filepath = path.join('book/media/image', req.file.filename)
-
-
-        console.log(req.body)
+        const coverFile = req.files.coverfile[0];
+        const readableFile = req.files.readablefile[0];
         
         const NewbookData = {
             ...req.body, 
-            cover_path: req.file.filename
+            ID_Author: Token.ID_Author,
+            cover_path : coverFile.filename || null,
+            readable_path : readableFile.filename || null
         };
 
         const oldBookData = await bookModel.getbookByIdDb(NewbookData.ID_Buku);
+
+        if(!oldBookData){
+            res.status(404).json({message: 'no book found'});
+            return
+        }
         
-        const oldimageFolderPath = path.join(process.cwd(), '/libralab-api/media/image/book', oldBookData.cover_path); // Your folder path
-        // Step 3: Check if the file exists
-        if (fs.existsSync(oldimageFolderPath)) {
-
-            // Step 4: Delete the file
-            await fs.promises.unlink(oldimageFolderPath); 
-            // Asynchronously delete the file
-            console.log(`Image deleted: ${oldimageFolderPath}`);
-        }
-        else {
-            res.status(404).json({message: 'no cover image found'});
-        }
-
         const result = await bookModel.putBookDb(NewbookData);
 
         if(!result){
@@ -142,8 +166,36 @@ export async function putBookByBookId(req,res){
             return
         }
 
-        res.status(200).json({
-            message: 'book information succesfully sent'})
+        const oldimageFolderPath = path.join(process.cwd(), '/libralab-api/media/image/book', oldBookData.cover_path); // Your folder path
+        // Step 3: Check if the file exists
+        if (fs.existsSync(oldimageFolderPath)) {
+
+            // Step 4: Delete the file
+            await fs.promises.unlink(oldimageFolderPath); 
+            // Asynchronously delete the file
+            console.log(`Image deleted: ${oldimageFolderPath}`);
+        }
+        else {
+            res.status(404).json({message: 'no cover image found'});
+            return
+        }
+        
+        const oldReadableFolderPath = path.join(process.cwd(), '/libralab-api/media/readable/book', oldBookData.readable_path); // Your folder path
+        // Step 3: Check if the file exists
+        if (fs.existsSync(oldReadableFolderPath)) {
+
+            // Step 4: Delete the file
+            await fs.promises.unlink(oldReadableFolderPath); 
+            // Asynchronously delete the file
+            console.log(`Readable deleted: ${oldReadableFolderPath}`);
+        }
+        else {
+            res.status(404).json({message: 'no readable found'});
+            return
+        }
+
+        res.status(200).json({message: 'book information succesfully sent'})
+            return
 
     } catch(error) {
         res.status(500).json({ 
@@ -174,6 +226,34 @@ export async function getBookCoverByUrl(req,res){
     } catch (error) {
         if(error.code === 'ENOENT'){
             res.status(404).json({ message: 'image not found' })
+            return
+        }
+        res.status(500).json({ message: 'Internal server error' })
+        console.log(error);
+    }
+}
+
+export async function getBookReadableByUrl(req,res){
+    try {
+        const {readable_path} = req.params;
+
+        if(!readable_path){
+            res.status(400).json({message:"no url specified"})
+            return
+        }
+
+        const filePath = path.join(process.cwd(), '/libralab-api/media/readable/book', readable_path)
+
+        if(!fs.existsSync(filePath)){
+            res.status(404).json({ message: 'readable not found' })
+            return
+        }
+
+        res.status(200).sendFile(filePath)
+        
+    } catch (error) {
+        if(error.code === 'ENOENT'){
+            res.status(404).json({ message: 'readable not found' })
             return
         }
         res.status(500).json({ message: 'Internal server error' })
@@ -259,7 +339,6 @@ export async function deleteBookById(req,res){
     try {
         const dataBuku = req.body
         const authHeader = req.cookies.jwt;
-        console.log(authHeader);
         const Token = await jwtMiddleware.isJWTValid(authHeader);
 
         if(Token === 403){
@@ -271,19 +350,11 @@ export async function deleteBookById(req,res){
             return
         } 
 
-        const { cover_path } = req.body;
+        const DeletedBookData = await bookModel.getbookByIdDb(dataBuku.ID_Buku)
 
-        const imageFolderPath = path.join(process.cwd(), '/libralab-api/media/image/book', cover_path); // Your folder path
-        // Step 3: Check if the file exists
-        if (fs.existsSync(imageFolderPath)) {
-
-            // Step 4: Delete the file
-            await fs.promises.unlink(imageFolderPath); 
-            // Asynchronously delete the file
-            console.log(`Image deleted: ${imageFolderPath}`);
-        }
-        else {
-            res.status(404).json({message: 'no cover image found'});
+        if(!DeletedBookData){
+            res.status(404).json({message: 'no Book in database for this author'});
+            return
         }
 
         const result = await bookModel.deleteBookDb(dataBuku.ID_Buku);
@@ -298,7 +369,40 @@ export async function deleteBookById(req,res){
             return
         }
 
-            res.status(200).json({message: 'Book successfully deleted'});
+        const { cover_path } = DeletedBookData;
+
+        const imageFolderPath = path.join(process.cwd(), '/libralab-api/media/image/book', cover_path); // Your folder path
+        // Step 3: Check if the file exists
+        if (fs.existsSync(imageFolderPath)) {
+
+            // Step 4: Delete the file
+            await fs.promises.unlink(imageFolderPath); 
+            // Asynchronously delete the file
+            console.log(`Image deleted: ${imageFolderPath}`);
+        }
+        else {
+            res.status(404).json({message: 'no cover image found'});
+            return;
+        }
+
+        const { readable_path } = DeletedBookData;
+
+        const oldReadableFolderPath = path.join(process.cwd(), '/libralab-api/media/readable/book', readable_path); // Your folder path
+        // Step 3: Check if the file exists
+        if (fs.existsSync(oldReadableFolderPath)) {
+
+            // Step 4: Delete the file
+            await fs.promises.unlink(oldReadableFolderPath); 
+            // Asynchronously delete the file
+            console.log(`Readable deleted: ${oldReadableFolderPath}`);
+        }
+        else {
+            res.status(404).json({message: 'no readable found'});
+            return
+        }
+
+
+        res.status(200).json({message: 'Book successfully deleted'});
 
     } catch (error) {
         res.status(500).json({ message: 'Internal server error' })
